requirements.txt
altgraph==0.17.2
async-generator==1.10
attrs==21.2.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.27
bcrypt==3.2.0
beautifulsoup4==4.9.3
cachetools==4.2.4
capstone==4.0.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.3
colorama==0.4.4
colored-traceback==0.3.0
cryptography==3.4.7
cycler==0.10.0
docopt==0.6.2
et-xmlfile==1.1.0
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
future==0.18.2
google-api-core==2.2.0
google-api-python-client==2.28.0
google-auth==2.3.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.1.0
google-cloud-datastore==2.4.0
google-cloud-storage==1.42.3
google-cloud-vision==2.6.1
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.41.1
grpcio-status==1.41.1
h11==0.12.0
httplib2==0.20.1
idna==2.10
install==1.3.4
intervaltree==3.1.0
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.1
libcst==0.3.23
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib==3.4.2
msrest==0.6.21
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector-python==8.0.27
mysqlclient==2.1.0
numpy==1.20.3
oauthlib==3.1.1
openpyxl==3.0.9
outcome==1.1.0
packaging==20.9
pandas==1.2.4
paramiko==2.7.2
pefile==2021.9.3
Pillow==8.4.0
pipreqs==0.4.11
plumbum==1.7.0
proto-plus==1.19.7
protobuf==3.19.0
psutil==5.8.0
pwntools==4.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyelftools==0.27
Pygments==2.9.0
pyinstaller==4.6
pyinstaller-hooks-contrib==2021.3
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==6.0
requests==2.25.1
requests-oauthlib==1.3.0
ROPGadget==6.5
rpyc==5.0.1
rsa==4.7.2
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
SQLAlchemy==1.4.27
trio==0.19.0
trio-websocket==0.9.2
typing-inspect==0.7.1
typing_extensions==4.0.1
unicorn==1.0.2rc3
uritemplate==4.1.1
urllib3==1.26.4
Werkzeug==2.0.2
win-inet-pton==1.1.0
wsproto==1.0.0
yarg==0.1.9
